requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozendict==2.4.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
importlib_metadata==8.4.0
iniconfig==2.0.0
Jinja2==3.1.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
numpy==2.0.1
opentelemetry-api==1.27.0
opentelemetry-exporter-zipkin==1.27.0
opentelemetry-exporter-zipkin-json==1.27.0
opentelemetry-exporter-zipkin-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-propagator-b3==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pluggy==1.5.0
protobuf==3.20.3
py_eureka_client==0.11.10
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyMySQL==1.1.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
yfinance==0.2.41
zipp==3.20.2
